refactor: bound MyBatis-Plus queries across support services
- replace unbounded list scans with shared maximum page limits and explicit query bounds across analytics and platform services

- move event-platform summary counts into repository aggregates and add repository behavior coverage

- remove duplicate module configuration classes now supplied by common auto-configuration

- delete source-string ModuleIntegrationTest placeholders from the migrated support modules

+++ event-platform/src/main/java/com/emall/eventplatform/EventPlatformDomain.java
@@ -32,6 +32,6 @@ record MetricMaterialization(long materializationId, String metricName, String w
         long lateEventCount, Instant materializedAt) {
 }
 
-record EventPlatformSummary(int activeSchemas, int ingestedEvents, int lateEvents, int materializedMetrics,
-        int classifiedFields) {
+record EventPlatformSummary(long activeSchemas, long ingestedEvents, long lateEvents, long materializedMetrics,
+        long classifiedFields) {
 }
